seperate anime grid render into one component
TrendingAnime, PopularAnime, FavouriteAnime, and Top100Anime, all have the exact jsx and styles with different values. Seperating them from the pages into a single component which takes props, allows for easier maintainability and accessibility.
